backend/requirements.txt
Django
djangorestframework
boto3
gunicorn
